ai_proofreading_tool/requirements.txt
PySide6==6.5.0
qfluentwidgets==1.2.0
openai==0.27.0
anthropic==0.2.6
google-cloud-aiplatform==1.25.0
requests==2.26.0
google-api-python-client==2.86.0
google-auth==2.19.1